+++ src/GraphQL.php
@@ -285,8 +285,12 @@ class GraphQL
         return $result->toArray($this->debugFlag());
     }
 
-    /** Loads persisted query from the query cache. */
-    protected function loadPersistedQuery(string $sha256hash): DocumentNode
+    /**
+     * Loads persisted query from the query cache.
+     *
+     * @api
+     */
+    public function loadPersistedQuery(string $sha256hash): DocumentNode
     {
         $lighthouseConfig = $this->configRepository->get('lighthouse');
         $cacheConfig = $lighthouseConfig['query_cache'] ?? null;