+++ CHANGELOG.md
@@ -9,6 +9,22 @@ You can find and compare releases at the [GitHub release page](https://github.co
 
 ## Unreleased
 
+## v6.40.0
+
+### Added
+
+- Make `GraphQL::loadPersistedQuery` public https://github.com/nuwave/lighthouse/pull/2579
+
+## v6.39.0
+
+### Added
+
+- Add directives `@createMany`, `@updateMany` and `@upsertMany` to mutate multiple models at once https://github.com/nuwave/lighthouse/pull/2578
+
+### Changed
+
+- Split `MutationExecutorDirective` into `OneModelMutationDirective` and `ManyModelMutationDirective` https://github.com/nuwave/lighthouse/pull/2578
+
 ## v6.38.1
 
 ### Fixed